disable compression in save step
The /nix/store can get very large. The default GitHub runners are very
slow to decompress and recompress large archives but the network
between the runner and the cache is very fast, so in order to optimize
the usecase for Nix further, simply disable compression outright and
rely on the network speed to make cache restore and upload fast.

This change is being split into two commits for a smooth transition for
those already using the previous tag. First point your action to this
commit, then to the subsequent commit on next run after a new cache with
compression disabled has been uploaded.